Update tests and locations to reflect cura module move
